connected to azure db!

+++ Models/LineItem.cs
@@ -7,9 +7,9 @@ public partial class LineItem
 {
     public int? TransactionId { get; set; }
 
-    public int? ProductId { get; set; }
+    public byte? ProductId { get; set; }
 
-    public int? Qty { get; set; }
+    public byte? Qty { get; set; }
 
-    public int? Rating { get; set; }
+    public byte? Rating { get; set; }
 }
