+++ Models/Product.cs
@@ -9,11 +9,11 @@ public partial class Product
 
     public string? Name { get; set; }
 
-    public int? Year { get; set; }
+    public short? Year { get; set; }
 
-    public int? NumParts { get; set; }
+    public short? NumParts { get; set; }
 
-    public int? Price { get; set; }
+    public short? Price { get; set; }
 
     public string? ImgLink { get; set; }
